feat(activitiesCompleted): isolate get completed activities controller
